requirements.txt
cssmin==0.2.0
dj-database-url==0.4.2
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.1
django-webpack-loader==0.4.1
djangorestframework==3.5.3
gunicorn==19.6.0
jsmin==2.2.1
nodeenv==1.0.0
psycopg2==2.6.2
rcssmin==1.0.6
rjsmin==1.0.12
whitenoise==3.2.3
